requirements.txt
fastapi==0.115.12
uvicorn==0.30.6
python-dotenv==1.0.1
